usando la variable de sesion

+++ app/Http/Controllers/PortfolioController.php
@@ -19,11 +19,15 @@ class PortfolioController extends Controller
      */
     public function index()
     {
-        $users = DB::table('users')->paginate(15);
-        return view('layout', [
-            'usuarios' => $users
-        ]);
-
+        if(session()->has('admin')==false){
+            return redirect('login');
+        }else{
+            
+            $users = DB::table('users')->paginate(15);
+                return view('layout', [
+                            'usuarios' => $users
+                ]);
+        }
         // return view('layout');
     }
 
@@ -89,59 +93,101 @@ class PortfolioController extends Controller
     }
 
     public function editar_datos($id) {
-        $valor = Project::findOrFail($id);
-        return view('informacion.edicion', compact('valor'));
+        if(session()->has('admin')==false){
+            return redirect('login');
+        }else{
+            
+            $valor = Project::findOrFail($id);
+            return view('informacion.edicion', compact('valor'));
+            
+        }
     }
 
     public function Editar_usuario(Request  $request)
     {
-        $id_usuario=$request->input('usuario');
-        $editar_usuario = DB::table('users')->where('ID_USER',$id_usuario)->first();
-        return view('Login.Registro',[
-            'editar_usu'=> $editar_usuario,
-            'id_usuario' => $id_usuario,
-        ]);
+        if(session()->has('admin')==false){
+            return redirect('login');
+        }else{
+            
+            $id_usuario=$request->input('usuario');
+            $editar_usuario = DB::table('users')->where('ID_USER',$id_usuario)->first();
+            return view('Login.Registro',[
+                'editar_usu'=> $editar_usuario,
+                'id_usuario' => $id_usuario,
+            ]);
+            
+        }
     }
 
     public function Eliminar($id_usuario)
     {
-        DB::table('users')->where('ID_USER', '=', $id_usuario)->delete();
-        return redirect("principal");
+        if(session()->has('admin')==false){
+            return redirect('login');
+        }else{
+            
+            DB::table('users')->where('ID_USER', '=', $id_usuario)->delete();
+            return redirect("principal");
+            
+        }
     }
 
     public function Elimiarimagen($id_imagen)
     {
-        DB::table('cv_imagenes')->where('IMAGENES_ID', '=', $id_imagen)->delete();
-        return redirect("VerPropiedades");
+        if(session()->has('admin')==false){
+            return redirect('login');
+        }else{
+            
+            DB::table('cv_imagenes')->where('IMAGENES_ID', '=', $id_imagen)->delete();
+            return redirect("VerPropiedades");
+            
+        }
     }
 
     public function Elimiarimageninicio( $id_imagen)
     {
-        DB::table('cv_inicio')->where('INICIO_ID', '=', $id_imagen)->delete();
-        return redirect("Verinicio");
+        if(session()->has('admin')==false){
+            return redirect('login');
+        }else{
+            
+            DB::table('cv_inicio')->where('INICIO_ID', '=', $id_imagen)->delete();
+            return redirect("Verinicio");
+            
+        }
     }
     public function guardarorden(Request  $request)
     {
-        $ordenarray=$request->input('orden');
-        for ($i=0; $i <count($ordenarray) ; $i++) { 
-            $orden = DB::table('cv_imagenes')
-                        ->where('IMAGENES_ID',$ordenarray[$i])
-                        ->update(['IMAGENES_ORDEN' => $i,]);
+        if(session()->has('admin')==false){
+            return redirect('login');
+        }else{
+            
+            $ordenarray=$request->input('orden');
+            for ($i=0; $i <count($ordenarray) ; $i++) { 
+                $orden = DB::table('cv_imagenes')
+                            ->where('IMAGENES_ID',$ordenarray[$i])
+                            ->update(['IMAGENES_ORDEN' => $i,]);
+            }
+            return back();
+            
         }
-        return back();
     }
 
     public function guradarmensages(contacto_validation $request)
     {
-        $contactos  = new cv_contactos;
-        $contactos->CONTACTO_NOMBRE     =   $request->nombre;
-        $contactos->CONTACTO_EMAIL      =   $request->email;
-        $contactos->CONTACTO_ASUNTO     =   $request->asunto;
-        $contactos->CONTACTO_TELEFONO   =   $request->tel;
-        $contactos->CONTACTO_MENSAJE    =   $request->mesaje;
-        $contactos->save();
-
-        return back();
+        if(session()->has('admin')==false){
+            return redirect('login');
+        }else{
+            
+            $contactos  = new cv_contactos;
+            $contactos->CONTACTO_NOMBRE     =   $request->nombre;
+            $contactos->CONTACTO_EMAIL      =   $request->email;
+            $contactos->CONTACTO_ASUNTO     =   $request->asunto;
+            $contactos->CONTACTO_TELEFONO   =   $request->tel;
+            $contactos->CONTACTO_MENSAJE    =   $request->mesaje;
+            $contactos->save();
+    
+            return back();
+            
+        }
     }
 
     public function Filtro_busquedad(Request  $request)
@@ -235,16 +281,28 @@ class PortfolioController extends Controller
     }
 
     public function VerContactos() {
-        $contactos = DB::table('cv_contactos')
-                    ->select('*')
-                    ->get();
-
-                    return view('administrador.ver_contactos', compact('contactos'));
+        if(session()->has('admin')==false){
+            return redirect('login');
+        }else{
+            
+            $contactos = DB::table('cv_contactos')
+                        ->select('*')
+                        ->get();
+    
+                        return view('administrador.ver_contactos', compact('contactos'));
+            
+        }
     }
 
     public function Eliminar_contacto($id_contacto)
     {
-        DB::table('cv_contactos')->where('ID_CONTACTO', '=', $id_contacto)->delete();
-        return redirect("VerContactos");
+        if(session()->has('admin')==false){
+            return redirect('login');
+        }else{
+            DB::table('cv_contactos')->where('ID_CONTACTO', '=', $id_contacto)->delete();
+            return redirect("VerContactos");
+            
+            
+        }
     }
 }
